Make edits to refactor
