+++ src/main/java/ListGPT/ListGPT.java
@@ -1,17 +1,54 @@
-package ListGPT;
+package Listgpt;
 
-public class ListGPT {
+import javafx.application.Application;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+public class Listgpt extends Application {
+
     private static TaskList list = new TaskList();
-    private static UI userInterface = new UI(list);
-    private static Storage storage = new Storage("./data/ListGPT.txt", list);
+    private static CLI messageInterface = new CLI(list);
+    private static GUI userInterface;
+    private static Storage storage = new Storage("./data/Listgpt.txt", list);
 
+    /**
+     * CLI Interface for ListGPT
+     * @param args
+     */
     public static void main(String[] args) {
         // Retrieve before running
         storage.retrieve();
 
-        userInterface.run();
+        messageInterface.run();
 
         // Save on exit
         storage.store();
+    }
+
+    /**
+     * GUI Interface for ListGPT
+     * @param stage the primary stage for this application, onto which
+     * the application scene can be set.
+     * Applications may create other stages, if needed, but they will not be
+     * primary stages.
+     */
+    @Override
+    public void start(Stage stage) {
+        userInterface = new GUI(list, messageInterface, storage);
+        Scene scene = userInterface.build();
+
+        stage.setTitle("ListGPT");
+        stage.setScene(scene);
+        stage.setResizable(true);
+        stage.setMinHeight(600.0);
+        stage.setMinWidth(400.0);
+        stage.setMaxHeight(800.0);
+        stage.setMaxWidth(600.0);
+        stage.show();
+    }
+
+    @Override
+    public void stop() {
+        if (userInterface != null) userInterface.onStop();
     }
 }
